refactor(tabs): extract shared tab-registry open helpers to lib
