+++ src/components/TabsView.tsx
@@ -1,43 +1,36 @@
 import { createElement, memo, useEffect, useMemo, useState } from 'react'
-import { nanoid } from 'nanoid'
 import {
   Archive,
-  Bot,
   ChevronDown,
   ChevronRight,
   Copy,
   ExternalLink,
-  FileCode2,
   Globe,
   Monitor,
-  Square,
-  TerminalSquare,
   type LucideIcon,
 } from 'lucide-react'
 import { useAppDispatch, useAppSelector, useAppStore } from '@/store/hooks'
 import { getWsClient } from '@/lib/ws-client'
-import { RegistryPaneSnapshotSchema, type RegistryPaneSnapshot, type RegistryTabRecord } from '@/store/tabRegistryTypes'
-import { addTab, setActiveTab } from '@/store/tabsSlice'
-import { addPane, initLayout } from '@/store/panesSlice'
+import type { RegistryPaneSnapshot, RegistryTabRecord } from '@/store/tabRegistryTypes'
 import { setTabRegistryLoading, setTabRegistrySearchRangeDays } from '@/store/tabRegistrySlice'
 import { selectTabsRegistryGroups } from '@/store/selectors/tabsRegistrySelectors'
 import { getCurrentTabRegistryClientInstanceId } from '@/store/tabRegistrySync'
-import { isNonShellMode } from '@/lib/coding-cli-utils'
 import { copyText } from '@/lib/clipboard'
 import { cn } from '@/lib/utils'
 import { ContextMenu } from '@/components/context-menu/ContextMenu'
 import type { MenuItem } from '@/components/context-menu/context-menu-types'
 import {
-  normalizeFreshAgentEffortOverride,
-  normalizeFreshAgentModelSelection,
-  type PaneContentInput,
-  type SessionLocator,
-} from '@/store/paneTypes'
-import type { CodingCliProviderName, TabMode } from '@/store/types'
-import { sanitizeRestoreError } from '@shared/session-contract'
-import { sanitizeCodexDurabilityRef } from '@shared/codex-durability'
-import { normalizeFreshAgentSessionType, resolveFreshAgentRuntimeProvider } from '@shared/fresh-agent'
-import { normalizeFreshAgentStyleOverride } from '@shared/settings'
+  jumpToRecord as jumpToRecordAction,
+  openPaneInNewTab as openPaneInNewTabAction,
+  openRecordAsUnlinkedCopy as openRecordAsUnlinkedCopyAction,
+  paneKindColorClass,
+  paneKindIcon,
+  paneKindLabel,
+} from '@/lib/tab-registry-open'
+
+// Compatibility re-export: test/unit/client/tab-registry-fresh-agent-migration.test.ts
+// (and any external caller) imports sanitizePaneSnapshot from this module.
+export { sanitizePaneSnapshot } from '@/lib/tab-registry-open'
 
 /* ------------------------------------------------------------------ */
 /*  Types                                                             */
@@ -57,195 +50,6 @@ type DeviceGroupData = {
 /* ------------------------------------------------------------------ */
 /*  Utilities (unchanged business logic)                              */
 /* ------------------------------------------------------------------ */
-
-function parseSessionLocator(value: unknown): SessionLocator | undefined {
-  if (!value || typeof value !== 'object') return undefined
-  const candidate = value as { provider?: unknown; sessionId?: unknown }
-  if (typeof candidate.provider !== 'string' || !isNonShellMode(candidate.provider)) {
-    return undefined
-  }
-  if (typeof candidate.sessionId !== 'string') return undefined
-  return {
-    provider: candidate.provider as CodingCliProviderName,
-    sessionId: candidate.sessionId,
-  }
-}
-
-function resolveSessionRef(options: {
-  payload: Record<string, unknown>
-  fallbackProvider?: CodingCliProviderName
-  fallbackSessionId?: string
-}): SessionLocator | undefined {
-  const explicit = parseSessionLocator(options.payload.sessionRef)
-  if (explicit) return explicit
-  if (!options.fallbackProvider || !options.fallbackSessionId) return undefined
-  return {
-    provider: options.fallbackProvider,
-    sessionId: options.fallbackSessionId,
-  }
-}
-
-function parseLiveTerminalHandle(
-  value: unknown,
-  recordServerInstanceId: string,
-): { terminalId: string; serverInstanceId: string } | undefined {
-  if (!value || typeof value !== 'object') return undefined
-  const candidate = value as { terminalId?: unknown; serverInstanceId?: unknown }
-  if (typeof candidate.terminalId !== 'string' || typeof candidate.serverInstanceId !== 'string') {
-    return undefined
-  }
-  if (candidate.serverInstanceId !== recordServerInstanceId) {
-    return undefined
-  }
-  return {
-    terminalId: candidate.terminalId,
-    serverInstanceId: candidate.serverInstanceId,
-  }
-}
-
-function normalizePaneSnapshot(snapshot: RegistryPaneSnapshot): RegistryPaneSnapshot {
-  const parsed = RegistryPaneSnapshotSchema.safeParse(snapshot)
-  return parsed.success ? parsed.data : snapshot
-}
-
-export function sanitizePaneSnapshot(
-  record: RegistryTabRecord,
-  rawSnapshot: RegistryPaneSnapshot,
-  localServerInstanceId?: string,
-): PaneContentInput {
-  const snapshot = normalizePaneSnapshot(rawSnapshot)
-  const payload = snapshot.payload || {}
-  const sameServer = !!localServerInstanceId && record.serverInstanceId === localServerInstanceId
-  if (snapshot.kind === 'terminal') {
-    const mode = (payload.mode as TabMode) || 'shell'
-    const sessionRef = resolveSessionRef({ payload })
-    const liveTerminal = parseLiveTerminalHandle(payload.liveTerminal, record.serverInstanceId)
-    const codexDurability = mode === 'codex'
-      ? sanitizeCodexDurabilityRef(payload.codexDurability)
-      : undefined
-    const includeLiveTerminal = sameServer && !sessionRef
-    return {
-      kind: 'terminal',
-      mode,
-      shell: (payload.shell as 'system' | 'cmd' | 'powershell' | 'wsl') || 'system',
-      sessionRef,
-      ...(codexDurability ? { codexDurability } : {}),
-      terminalId: includeLiveTerminal ? liveTerminal?.terminalId : undefined,
-      serverInstanceId: includeLiveTerminal ? record.serverInstanceId : undefined,
-      initialCwd: payload.initialCwd as string | undefined,
-    }
-  }
-  if (snapshot.kind === 'browser') {
-    return {
-      kind: 'browser',
-      url: (payload.url as string) || 'https://example.com',
-      devToolsOpen: !!payload.devToolsOpen,
-    }
-  }
-  if (snapshot.kind === 'editor') {
-    return {
-      kind: 'editor',
-      filePath: (payload.filePath as string | null) ?? null,
-      language: (payload.language as string | null) ?? null,
-      readOnly: !!payload.readOnly,
-      content: '',
-      viewMode: (payload.viewMode as 'source' | 'preview') || 'source',
-      wordWrap: payload.wordWrap !== false,
-    }
-  }
-  if (snapshot.kind === 'fresh-agent') {
-    const sessionType = normalizeFreshAgentSessionType(payload.sessionType)
-      ?? normalizeFreshAgentSessionType(payload.provider)
-    const provider = (
-      payload.provider === 'claude'
-      || payload.provider === 'codex'
-      || payload.provider === 'opencode'
-    )
-      ? payload.provider
-      : resolveFreshAgentRuntimeProvider(sessionType)
-    if (!sessionType || !provider) return { kind: 'picker' }
-    const resumeSessionId = typeof payload.resumeSessionId === 'string'
-      ? payload.resumeSessionId
-      : undefined
-    const sessionRef = resolveSessionRef({
-      payload,
-      fallbackProvider: provider,
-      fallbackSessionId: resumeSessionId,
-    })
-    const style = normalizeFreshAgentStyleOverride(payload.style)
-    const restoreError = sanitizeRestoreError(payload.restoreError)
-    return {
-      kind: 'fresh-agent',
-      sessionType,
-      provider,
-      resumeSessionId,
-      ...(sessionRef ? { sessionRef } : {}),
-      ...(restoreError && !sessionRef ? { restoreError } : {}),
-      serverInstanceId: record.serverInstanceId,
-      initialCwd: payload.initialCwd as string | undefined,
-      model: payload.model as string | undefined,
-      modelSelection: normalizeFreshAgentModelSelection(payload.modelSelection, payload.model),
-      permissionMode: payload.permissionMode as string | undefined,
-      sandbox: payload.sandbox as 'read-only' | 'workspace-write' | 'danger-full-access' | undefined,
-      effort: normalizeFreshAgentEffortOverride(payload.effort),
-      plugins: payload.plugins as string[] | undefined,
-      ...(style ? { style } : {}),
-      settingsDismissed: typeof payload.settingsDismissed === 'boolean' ? payload.settingsDismissed : undefined,
-      showThinking: typeof payload.showThinking === 'boolean' ? payload.showThinking : undefined,
-      showTools: typeof payload.showTools === 'boolean' ? payload.showTools : undefined,
-      showTimecodes: typeof payload.showTimecodes === 'boolean' ? payload.showTimecodes : undefined,
-    }
-  }
-  if (snapshot.kind === 'extension') {
-    return {
-      kind: 'extension',
-      extensionName: (payload.extensionName as string) || 'unknown',
-      props: (payload.props as Record<string, unknown>) || {},
-    }
-  }
-  return { kind: 'picker' }
-}
-
-function deriveModeFromRecord(record: RegistryTabRecord): TabMode {
-  const firstKind = record.panes[0]?.kind
-  if (firstKind === 'terminal') {
-    const mode = record.panes[0]?.payload?.mode
-    if (typeof mode === 'string') return mode as TabMode
-    return 'shell'
-  }
-  if (firstKind === 'fresh-agent') {
-    const provider = record.panes[0]?.payload?.provider
-    if (typeof provider === 'string' && isNonShellMode(provider)) return provider as TabMode
-    return 'claude'
-  }
-  return 'shell'
-}
-
-function paneKindIcon(kind: RegistryPaneSnapshot['kind']): LucideIcon {
-  if (kind === 'terminal') return TerminalSquare
-  if (kind === 'browser') return Globe
-  if (kind === 'editor') return FileCode2
-  if (kind === 'fresh-agent') return Bot
-  return Square
-}
-
-function paneKindColorClass(kind: RegistryPaneSnapshot['kind']): string {
-  if (kind === 'terminal') return 'text-foreground/50'
-  if (kind === 'browser') return 'text-blue-500'
-  if (kind === 'editor') return 'text-emerald-500'
-  if (kind === 'fresh-agent' || kind === 'claude-chat') return 'text-amber-500'
-  if (kind === 'extension') return 'text-purple-500'
-  return 'text-muted-foreground'
-}
-
-function paneKindLabel(kind: RegistryPaneSnapshot['kind']): string {
-  if (kind === 'terminal') return 'Terminal'
-  if (kind === 'browser') return 'Browser'
-  if (kind === 'editor') return 'Editor'
-  if (kind === 'fresh-agent' || kind === 'claude-chat') return 'Agent'
-  if (kind === 'extension') return 'Extension'
-  return kind
-}
 
 function formatRelativeTime(timestamp: number, now: number): string {
   const diff = Math.max(0, now - timestamp)
@@ -601,56 +405,21 @@ function TabsView({ onOpenTab }: { onOpenTab?: () => void }) {
   /* -- actions ------------------------------------------------------ */
 
   const openRecordAsUnlinkedCopy = (record: RegistryTabRecord) => {
-    const tabId = nanoid()
-    const paneSnapshots = record.panes || []
-    const firstPane = paneSnapshots[0]
-    const firstContent = firstPane
-      ? sanitizePaneSnapshot(record, firstPane, localServerInstanceId)
-      : ({ kind: 'terminal', mode: 'shell' } as const)
-    dispatch(
-      addTab({
-        id: tabId,
-        title: record.tabName,
-        mode: deriveModeFromRecord(record),
-        status: 'creating',
-        serverInstanceId: record.serverInstanceId,
-      }),
-    )
-    dispatch(initLayout({ tabId, content: firstContent }))
-    for (const pane of paneSnapshots.slice(1)) {
-      dispatch(addPane({ tabId, newContent: sanitizePaneSnapshot(record, pane, localServerInstanceId) }))
-    }
-    onOpenTab?.()
+    openRecordAsUnlinkedCopyAction(record, { dispatch, localServerInstanceId, onOpened: onOpenTab })
   }
 
+  // Still used by openCardContextMenu until Task 6 removes the local menu.
   const openPaneInNewTab = (record: RegistryTabRecord, pane: RegistryPaneSnapshot) => {
-    const tabId = nanoid()
-    dispatch(
-      addTab({
-        id: tabId,
-        title: `${record.tabName} · ${pane.title || pane.kind}`,
-        mode: deriveModeFromRecord(record),
-        status: 'creating',
-        serverInstanceId: record.serverInstanceId,
-      }),
-    )
-    dispatch(
-      initLayout({
-        tabId,
-        content: sanitizePaneSnapshot(record, pane, localServerInstanceId),
-      }),
-    )
-    onOpenTab?.()
+    openPaneInNewTabAction(record, pane, { dispatch, localServerInstanceId, onOpened: onOpenTab })
   }
 
   const jumpToRecord = (record: RegistryTabRecord) => {
-    const localTabExists = store.getState().tabs.tabs.some((tab) => tab.id === record.tabId)
-    if (!localTabExists) {
-      openRecordAsUnlinkedCopy(record)
-      return
-    }
-    dispatch(setActiveTab(record.tabId))
-    onOpenTab?.()
+    jumpToRecordAction(record, {
+      dispatch,
+      localServerInstanceId,
+      onOpened: onOpenTab,
+      hasLocalTab: (tabId) => store.getState().tabs.tabs.some((tab) => tab.id === tabId),
+    })
   }
 
   const pullAllFromDevice = (tabs: DisplayRecord[]) => {
